+++ examples/M1/M1_run.py
@@ -3,5 +3,5 @@ from netpyne import sim  # import netpyne init module
 
 sim.createSimulateAnalyze(netParams = M1.netParams, simConfig = M1.simConfig)  # create and simulate network
 
-# check model output
+# Check the model output: sim.checkOutput is used for testing purposes.  Please comment out the following line if you are exploring the example.
 sim.checkOutput('M1')